fix(Selector): moved 'Selector' class to it's own module.
